+++ tests/supportingResources/versions.tf
@@ -2,7 +2,11 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = ">=4.15.1"
+      version = ">= 5.0.0, < 6.0.0"
+    }
+    random = {
+      source  = "hashicorp/random"
+      version = ">= 3.0.0"
     }
   }
   required_version = ">= 0.14.11"